+++ tests/error_messages.rs
@@ -47,7 +47,7 @@ foo,bar,[[text=bold ,foo
     assert_eq!(
         module.unwrap_err().to_string(),
         "Syntax error in cell C1 of integration_test_syntax_error_in_option_definition.csvpp
-On line 3 Error parsing input, expected ']]' but saw unrecognized token ``
+On line 3 Expected a OptionName but saw unrecognized token ``
 
  1: 
  2: ---
@@ -91,7 +91,7 @@ fn syntax_error_in_csv_section() {
         "syntax_error_in_csv_section",
         "csv",
         r#"
-# it's a common problem that the `function_in_file1(1, 2)` call needs to be quoted. 
+# it's a common problem that the `function_in_file1(1, 2)` call needs to be quoted 
 # because it has a comma
 ---
 function_in_file1(1 * 2)  ,   =function_in_file1(1, 2)    , should be 1 * 44
@@ -103,7 +103,7 @@ function_in_file1(1 * 2)  ,   =function_in_file1(1, 2)    , should be 1 * 44
 On line 5 Expected an expression but saw EOF
 If your formula has a comma in it, you might need to escape it with quotes (i.e. `foo,\"=my_function(1, 2)\",bar`)
 
- 2: # it's a common problem that the `function_in_file1(1, 2)` call needs to be quoted. 
+ 2: # it's a common problem that the `function_in_file1(1, 2)` call needs to be quoted 
  3: # because it has a comma
  4: ---
  5: function_in_file1(1 * 2)  ,   =function_in_file1(1, 2)    , should be 1 * 44